fix: isolate model context formatter
